src/requirements.txt
langchain 
streamlit
fastapi
uvicorn
groq 
langchain-groq
python-dotenv